# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Solar Projects API",
        "description": "API for managing solar projects",
        "version": "1.0"
    },
    "paths": {
...
        "/api/customers": {
            "post": {
                "tags": [
                    "Customers"
                ],
                "summary": "Create a new customer",
                "operationId": "747630c757eae4c38404384330e68c57",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Customer"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Customer created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Customer"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/customers/{id}": {
            "get": {
                "tags": [
                    "Customers"
                ],
                "summary": "Display the specified customer",
                "operationId": "d03d1278b7e26322e9a425406bc4e9f2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the customer",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Customer details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Customer"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Customers"
                ],
                "summary": "Update the specified customer",
                "operationId": "32da1527afa82d1a6b0e034146f1032d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the customer",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Customer"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Customer updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Customer"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Customers"
                ],
                "summary": "Remove the specified customer",
                "operationId": "e9f16ac94d63309225711a1f3dfb245a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the customer",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Customer deleted successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Customer": {
                "required": [
                    "name",
                    "email",
                    "phone",
                    "document"
                ],
                "properties": {
                    "name": {
                        "description": "Name of the customer",
                        "type": "string",
                        "example": "Customer Test"
                    },
                    "email": {
                        "description": "Email of the customer",
                        "type": "string",
                        "format": "email",
                        "example": "customer.john@example.com"
                    },
                    "phone": {
                        "description": "Phone number of the customer",
                        "type": "string",
                        "example": "(82) 3561-3037"
                    },
                    "document": {
                        "description": "Document of the customer",
                        "type": "string",
                        "example": "140.873.430-30"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Bearer Token authentication. Enter token in format (Bearer <token>)",
                "scheme": "bearer"
            }
        }
    },
...
}